Initial commit
This has a sample job returned by the server right now and just logs
all traces (job log output). Next step is to hook this into webhooks
from GitHub, probably the runner to fetch a .gitlab-ci.yml file out of
the repository as an artifact, and then schedule jobs from there.

This includes code from gitlab-runner, and while I've tried to treat
that appropriately, changes may need to be made to make sure it's
handled correctly.