Initial commit of sqlite DB package
In the interest of getting slightly faster iteration going on this, the
first database / storage backend will be sqlite, thanks heavily to
David Crawshaw's work on his sqlite package (because the alternative is
using database/sql, which is really cumbersome).

So far this is just establishing some basic code to handle applying
database version patches to the sqlite file (or memory). It'll be
necessary to add a more general interface for creating and looking up
runners, taking a job, updating that job, etc. As a result, there isn't
really a reason to define a schema or cross-DB migrations or anything
that's shared across all storage backends. They'll just manage their
own patches or what have you.