Add basic migrations and initial table definitions
This first patch will change while I'm writing this, so there's no
reason to even assume migrations of any kind will be written in full
yet.

Right now this is just establishing a runners table and a base job
table (I know it needs an ID, I haven't determined what else should be
stored in-database versus memory versus filesystem). The point is
mostly that it's past 11pm right now and I want this committed. This
branch will probably continue to be force-pushed to.