Add token auth to RegisterRunner
This is to get the early part of authentication out of the way.
I'm not sure that GitLab even bothers with hashing their token, or how
it's verified when a runner is added, so I should probably check that
at some point. For now, the token is stored as a bcrypt hash and will
later be available for viewing through another authenticated request
(not using the token itself since that defeats the purpose). Maybe an
admin socket.