Implement basic runner registration
This now allows a runner to register itself with gribble. No
authentication on runner paths is handled yet, so that's necessary
later.

I'm not convinced the runner tagging setup works correctly yet, but
given that changing tags is unlikely and I typically just want to get
a runner with all of a set of tags, this'll probably work for now. Can
refactor it later if I think of something simpler / clearly better.

- Add a CreateRunner method to the DB interface. There's an additional
  TagRunner method on SQLite that should be useful for updating runner
  info from an administrative point, but it doesn't look like GitLab
  runners have any ability to modify their own information after
  sending it.

- Add genToken function that pulls some stuff out of randomToken. This
  is used to generate runner tokens.

- Add DB field to Server so that it can interact with things.
  Pondering whether the server should have the DB in it, but for now,
  this gets things rolling.

- Add new common/ package. This'll likely hold the types shared between
  gribblesv and database implementations.