Add runner lookup methods to DB interface
This is necessary so that when a runner requests a job, we can fetch
the runner's status, see if it's active, locked, whether any jobs match
any particular tags, etc.

Not really a lot of meaningful code here, just cleaning up some things
(e.g., uses of PrepareTransient that can be ExecTransient).