Add timestamps for runners
This is more important for jobs, but since I needed to decide on time
encoding in sqlite (float64 for fractinal seconds seemed reasonable),
times have been added to runners.

The runner updated time is essentially the last time the runner
requested a job, was tagged, was deactivated, etc.

This adds a proc package that can be used to override system time by
context.