Add skeleton webhook handler to Server
Adds a GitHub webhook handler to the Server so that these can be
received. It only logs the webhook, if successfuly parsed, if one is
received. No project provisioning or anything else is available yet.
There is also no validation of the payloads because I'm testing with
a sample payload still.