Replace stdlib log with zap
This is so that I can get this in now before it becomes a burden to
replace later, since I have no plans to use the stdlib log package for
Gribble by the time it's done.

Zap works well for JSON / console logging with structured data, so
that's what I'm going with. Encoder settings are intended to make
inputs human-readable even if they're JSON.